Refactor and rename service management system.
