+++ service/src/main/java/org/akazukin/service/data/BlueprintedServiceHolder.java
@@ -5,11 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.experimental.FieldDefaults;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.Objects;
 
 /**
- * A concrete implementation of the {@link IServiceHolder} interface.
+ * A concrete implementation of the {@link IBlueprintedServiceHolder} interface.
  * This class serves as a type-safe holder for managing both the interface
  * class and the implementation of a specific service.
  *
@@ -18,11 +19,12 @@ import java.util.Objects;
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 @Getter
-public final class SingleServiceHolder<T> implements ISingleServiceHolder<T> {
+public final class BlueprintedServiceHolder<T> implements IBlueprintedServiceHolder<T> {
+    @Nullable Class<T> interfaceClass;
     @NotNull T implementation;
 
     @Override
     public int hashCode() {
-        return Objects.hashCode(this.implementation.getClass());
+        return Objects.hashCode(this.interfaceClass);
     }
 }
