+++ service/src/main/java/org/akazukin/service/data/CompoundServiceHolder.java
@@ -12,7 +12,7 @@ import org.jetbrains.annotations.Nullable;
 import java.util.Objects;
 
 /**
- * A concrete implementation of the {@link IServiceHolder} interface.
+ * A concrete implementation of the {@link IBlueprintedServiceHolder} interface.
  * This class serves as a type-safe holder for managing both the interface
  * class and the implementation of a specific service.
  *
@@ -23,8 +23,6 @@ import java.util.Objects;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Getter
 public class CompoundServiceHolder<T, U> implements ICompoundServiceHolder<T, U> {
-    @Nullable
-    final Class<T> interfaceClass;
     @NotNull
     final T implementation;
     @Setter
@@ -32,6 +30,6 @@ public class CompoundServiceHolder<T, U> implements ICompoundServiceHolder<T, U>
 
     @Override
     public int hashCode() {
-        return Objects.hashCode(this.interfaceClass);
+        return Objects.hashCode(this.implementation.getClass());
     }
 }
