+++ service/src/main/java/org/akazukin/service/data/IServiceHolder.java
@@ -1,24 +1,23 @@
 package org.akazukin.service.data;
 
-import org.jetbrains.annotations.Nullable;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * Represents a holder interface for managing a service of a specific type.
- * This interface is designed to provide type-safe access to both the interface
- * class and the implementation of a service.
+ * This interface is designed to provide type-safe access to the implementation of a service.
  *
  * @param <T> the type of the service
  */
-public interface IServiceHolder<T> extends ISingleServiceHolder<T> {
+public interface IServiceHolder<T> {
     /**
-     * Retrieves the interface class associated with the service type managed by the holder.
-     * This method returns the class object representing the service interface,
-     * which may be useful for identifying or retrieving services
-     * based on their interface types.
+     * Retrieves the concrete implementation of the service managed by the holder.
+     * This method provides access to the actual service instance,
+     * which is guaranteed to be not {@code null}
+     * and can be used to perform operations specific to the service type.
      *
-     * @return the class object of the service interface,
-     * or {@code null} if the service is not explicitly associated with an interface class
+     * @return the implementation of the service managed by the holder
+     * Must not be {@code null}.
      */
-    @Nullable
-    Class<T> getInterfaceClass();
+    @NotNull
+    T getImplementation();
 }
