+++ service/src/main/java/org/akazukin/service/data/ServiceHolder.java
@@ -5,12 +5,11 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.experimental.FieldDefaults;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.Objects;
 
 /**
- * A concrete implementation of the {@link IServiceHolder} interface.
+ * A concrete implementation of the {@link IBlueprintedServiceHolder} interface.
  * This class serves as a type-safe holder for managing both the interface
  * class and the implementation of a specific service.
  *
@@ -20,11 +19,10 @@ import java.util.Objects;
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 @Getter
 public final class ServiceHolder<T> implements IServiceHolder<T> {
-    @Nullable Class<T> interfaceClass;
     @NotNull T implementation;
 
     @Override
     public int hashCode() {
-        return Objects.hashCode(this.interfaceClass);
+        return Objects.hashCode(this.implementation.getClass());
     }
 }
