+++ service/src/main/java/org/akazukin/service/manager/ABlueprintedCompoundServiceManager.java
@@ -3,7 +3,7 @@ package org.akazukin.service.manager;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
 import org.akazukin.annotation.marker.ThreadSafe;
-import org.akazukin.service.data.ICompoundSingleServiceHolder;
+import org.akazukin.service.data.IBlueprintedCompoundServiceHolder;
 import org.akazukin.util.utils.ArrayUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -12,7 +12,7 @@ import java.util.Objects;
 
 /**
  * An abstract implementation of a compound service manager that manages services and their associated data.
- * Extends the {@link AServiceManager} with additional functionalities for handling data linked with service holders.
+ * Extends the {@link ABlueprintedServiceManager} with additional functionalities for handling data linked with service holders.
  * <p>
  * The service manager is thread-safe and can be used in multithreaded environments.
  *
@@ -21,8 +21,8 @@ import java.util.Objects;
  */
 @FieldDefaults(level = AccessLevel.PROTECTED, makeFinal = true)
 @ThreadSafe
-public abstract class ACompoundSingleServiceManager<T extends ICompoundSingleServiceHolder<? extends U, V>, U, V>
-        extends ASingleServiceManager<T, U> implements ICompoundSingleServiceManager<T, U, V> {
+public abstract class ABlueprintedCompoundServiceManager<T extends IBlueprintedCompoundServiceHolder<? extends U, V>, U, V>
+        extends ABlueprintedServiceManager<T, U> implements IBlueprintedCompoundServiceManager<T, U, V> {
     Class<V> dataType;
 
     /**
@@ -37,7 +37,7 @@ public abstract class ACompoundSingleServiceManager<T extends ICompoundSingleSer
      *                          Must not be null.
      * @param dataType          the class object representing the type of data associated with the services.
      */
-    public ACompoundSingleServiceManager(@NotNull final Class<T> serviceHolderType, final @NotNull Class<U> serviceType, final Class<V> dataType) {
+    public ABlueprintedCompoundServiceManager(@NotNull final Class<T> serviceHolderType, final @NotNull Class<U> serviceType, final Class<V> dataType) {
         super(serviceHolderType, serviceType);
         this.dataType = dataType;
     }
@@ -47,7 +47,7 @@ public abstract class ACompoundSingleServiceManager<T extends ICompoundSingleSer
         return this.services.stream()
                 .filter(s -> Objects.equals(s.getImplementation().getClass(), service))
                 .findFirst()
-                .map(ICompoundSingleServiceHolder::getData)
+                .map(IBlueprintedCompoundServiceHolder::getData)
                 .orElse(null);
     }
 
@@ -56,14 +56,14 @@ public abstract class ACompoundSingleServiceManager<T extends ICompoundSingleSer
         return this.services.stream()
                 .filter(s -> s.getImplementation() == service)
                 .findFirst()
-                .map(ICompoundSingleServiceHolder::getData)
+                .map(IBlueprintedCompoundServiceHolder::getData)
                 .orElse(null);
     }
 
     @Override
     public V[] getAllData() {
         return this.services.stream()
-                .map(ICompoundSingleServiceHolder::getData)
+                .map(IBlueprintedCompoundServiceHolder::getData)
                 .toArray(ArrayUtils.collectToArray(this.dataType));
     }
 
@@ -72,5 +72,14 @@ public abstract class ACompoundSingleServiceManager<T extends ICompoundSingleSer
         return this.services.stream()
                 .filter(s -> Objects.equals(s.getData(), data))
                 .toArray(ArrayUtils.collectToArray(this.serviceHolderType));
+    }
+
+    @Override
+    public V getDataByInterface(final Class<? extends U> service) {
+        return this.services.stream()
+                .filter(s -> Objects.equals(s.getInterfaceClass(), service))
+                .findFirst()
+                .map(IBlueprintedCompoundServiceHolder::getData)
+                .orElse(null);
     }
 }
