+++ service/src/main/java/org/akazukin/service/manager/ServiceManager.java
@@ -1,23 +1,23 @@
 package org.akazukin.service.manager;
 
 import org.akazukin.annotation.marker.ThreadSafe;
+import org.akazukin.service.data.IBlueprintedServiceHolder;
 import org.akazukin.service.data.IServiceHolder;
-import org.akazukin.service.data.ServiceHolder;
+import org.akazukin.service.data.BlueprintedServiceHolder;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 /**
- * An implementation of the {@link AServiceManager} class for managing services
- * with {@link IServiceHolder} as the holder type.
+ * An implementation of the {@link ASingleServiceManager} class for managing services
+ * with {@link org.akazukin.service.data.IServiceHolder} as the holder type.
  * This class simplifies the process of creating service holders by using a specified type.
  *
  * @param <T> The type of the service being managed by this ServiceManager.
  */
 @ThreadSafe
-public class ServiceManager<T> extends AServiceManager<IServiceHolder<? extends T>, T> {
+public class ServiceManager<T> extends ASingleServiceManager<IServiceHolder<? extends T>, T> {
     /**
      * Constructs a {@link ServiceManager} instance for managing services of the specified type.
-     * This constructor leverages the {@link IServiceHolder} class for service holder management.
+     * This constructor leverages the {@link org.akazukin.service.data.IServiceHolder} class for service holder management.
      *
      * @param serviceType The class object representing the type of the service to be managed.
      *                    Must not be {@code null}.
@@ -28,7 +28,7 @@ public class ServiceManager<T> extends AServiceManager<IServiceHolder<? extends 
     }
 
     @Override
-    protected <T2 extends T> @NotNull IServiceHolder<? extends T> createServiceHolder(final @Nullable Class<T2> service, @NotNull final T2 serviceImpl) {
-        return new ServiceHolder<>(service, serviceImpl);
+    protected <T2 extends T> @NotNull IBlueprintedServiceHolder<? extends T> createServiceHolder(@NotNull final T2 serviceImpl) {
+        return new BlueprintedServiceHolder<>(null, serviceImpl);
     }
 }
